Remove need for VBP file completely. Now builds by default with default stuff.

+++ src/Cogito.VisualBasic6.MSBuild/VB6C.cs
@@ -19,6 +19,8 @@ namespace Cogito.VisualBasic6.MSBuild.Tasks
         Microsoft.Build.Utilities.Task
     {
 
+        VB6Type type;
+
         /// <summary>
         /// Path to the VB6 executable.
         /// </summary>
@@ -29,6 +31,21 @@ namespace Cogito.VisualBasic6.MSBuild.Tasks
         /// Source VBP project file to use as a base.
         /// </summary>
         public string Source { get; set; }
+
+        /// <summary>
+        /// Type of the project.
+        /// </summary>
+        [Required]
+        public string Type
+        {
+            get => Enum.GetName(typeof(VB6Type), type);
+            set => type = (VB6Type)Enum.Parse(typeof(VB6Type), value);
+        }
+
+        /// <summary>
+        /// Type of startup.
+        /// </summary>
+        public string Startup { get; set; }
 
         /// <summary>
         /// References to include in the VBP.
@@ -62,12 +79,20 @@ namespace Cogito.VisualBasic6.MSBuild.Tasks
         public string Output { get; set; }
 
         /// <summary>
+        /// Whether temporary files should be preserved.
+        /// </summary>
+        public bool PreserveTemporary { get; set; }
+
+        /// <summary>
         /// Applies the transforms.
         /// </summary>
         /// <param name="project"></param>
         /// <returns></returns>
         void Apply(VB6Project project)
         {
+            project.Type = type;
+            project.Startup = Startup;
+
             if (References != null &&
                 References.Length > 0)
             {
@@ -281,8 +306,8 @@ namespace Cogito.VisualBasic6.MSBuild.Tasks
         {
             // setup VB6 project
             var log = new StringWriter();
-            var src = Source != null ? VB6Project.Load(Source) : new VB6Project();
-            Apply(src);
+            var source = Source != null ? VB6Project.Load(Source) : new VB6Project();
+            Apply(source);
 
             // output to random directory to prevent it from deleting contents of original
             var dst = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
@@ -293,15 +318,12 @@ namespace Cogito.VisualBasic6.MSBuild.Tasks
 
                 using (var mutex = new Mutex(true, new Guid(MD5.Create().ComputeHash(Encoding.UTF8.GetBytes(Output))).ToString("n")))
                 {
-                    var errors = Task.Run(() =>
-                            new Compiler().Compile(
-                                new FileInfo(ToolPath),
-                                src,
-                                new DirectoryInfo(Output)))
-                        .GetAwaiter()
-                        .GetResult();
+                    var compiler = new Compiler(ToolPath)
+                    {
+                        PreserveTemporary = PreserveTemporary
+                    };
 
-                    if (errors != null)
+                    if (compiler.Compile(source, Output, out var errors) == false)
                     {
                         foreach (var error in errors)
                             Log.LogError(error);
@@ -322,8 +344,9 @@ namespace Cogito.VisualBasic6.MSBuild.Tasks
             {
                 try
                 {
-                    if (Directory.Exists(dst))
-                        Directory.Delete(dst, true);
+                    if (PreserveTemporary == false)
+                        if (Directory.Exists(dst))
+                            Directory.Delete(dst, true);
                 }
                 catch
                 {
