+++ src/Cogito.VisualBasic6.VB6C/Compiler.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+
 using Cogito.VisualBasic6.VB6C.EasyHook;
 using Cogito.VisualBasic6.VB6C.Project;
 
@@ -16,20 +17,47 @@ namespace Cogito.VisualBasic6.VB6C
     {
 
         /// <summary>
-        /// Compiles a <see cref="VB6Project"/>.
+        /// Initializes a new instnace.
+        /// </summary>
+        /// <param name="vb6exe"></param>
+        public Compiler(string vb6exe)
+        {
+            if (string.IsNullOrWhiteSpace(vb6exe))
+                throw new ArgumentException("Invalid VB6 executable path.", nameof(vb6exe));
+            if (File.Exists(vb6exe) == false)
+                throw new FileNotFoundException("Missing VB6 executable.");
+
+            VB6Exe = vb6exe;
+        }
+
+        /// <summary>
+        /// Path to the VB6 EXE.
+        /// </summary>
+        public string VB6Exe { get; }
+
+        /// <summary>
+        /// Preserves temporary files generated during the compilation.
+        /// </summary>
+        public bool PreserveTemporary { get; set; }
+
+        /// <summary>
+        /// Begins compilation. Returns <c>true</c> for successful compilation.
         /// </summary>
         /// <param name="vb6"></param>
         /// <param name="project"></param>
         /// <param name="output"></param>
-        /// <param name="logger"></param>
-        public List<string> Compile(FileInfo vb6, VB6Project project, DirectoryInfo output)
+        /// <param name="errors"></param>
+        /// <returns></returns>
+        public bool Compile(VB6Project project, string output, out IList<string> errors)
         {
-            if (vb6 == null)
-                throw new ArgumentNullException(nameof(vb6));
             if (project == null)
                 throw new ArgumentNullException(nameof(project));
             if (output == null)
                 throw new ArgumentNullException(nameof(output));
+
+            // create missing directory
+            if (Directory.Exists(output) == false)
+                Directory.CreateDirectory(output);
 
             var source = Path.Combine(Path.GetTempPath(), Path.ChangeExtension(Path.GetRandomFileName(), ".vbp"));
 
@@ -39,28 +67,33 @@ namespace Cogito.VisualBasic6.VB6C
 
                 var sti = new ProcessStartInfo();
                 sti.FileName = typeof(Executor).Assembly.Location;
-                sti.Arguments = $@"-e ""{vb6}"" -v ""{source}"" -o ""{output.FullName}""";
+                sti.Arguments = $@"-e ""{VB6Exe}"" -v ""{source}"" -o ""{output}""";
                 sti.UseShellExecute = false;
                 sti.RedirectStandardOutput = true;
                 sti.RedirectStandardError = true;
 
                 using (var prc = Process.Start(sti))
                 {
-                    var errors = new List<string>();
-                    prc.ErrorDataReceived += (s, a) => errors.Add(a.Data);
+                    var stderr = new List<string>();
+                    prc.ErrorDataReceived += (s, a) => stderr.Add(a.Data);
                     prc.BeginErrorReadLine();
                     prc.BeginOutputReadLine();
                     prc.WaitForExit();
+
                     if (prc.ExitCode != 0)
-                        return errors.Select(i => i?.Trim()).Where(i => !string.IsNullOrWhiteSpace(i)).ToList();
+                    {
+                        errors = stderr.Select(i => i?.Trim()).Where(i => !string.IsNullOrWhiteSpace(i)).ToList();
+                        return false;
+                    }
                 }
             }
             finally
             {
                 try
                 {
-                    if (File.Exists(source))
-                        File.Delete(source);
+                    if (PreserveTemporary == false)
+                        if (File.Exists(source))
+                            File.Delete(source);
                 }
                 catch
                 {
@@ -68,7 +101,8 @@ namespace Cogito.VisualBasic6.VB6C
                 }
             }
 
-            return null;
+            errors = null;
+            return true;
         }
 
     }
